requirements.txt
gradio>=4.0.0
together>=1.0.0
python-dotenv>=1.0.0
